requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-extensions==3.2.1
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-yasg==1.21.4
flake8==6.0.0
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
localStoragePy==0.2.3
Markdown==3.4.1
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
Pillow==9.4.0
pycodestyle==2.10.0
pydot==1.4.2
pyflakes==3.0.1
pyparsing==3.0.9
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.12.1
